monster/requirements.txt
asgiref==3.8.0
asttokens==2.4.1
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
Django==5.0.3
django-annoying==0.10.6
django-extensions==3.2.3
djangorestframework==3.15.0
executing==2.0.1
filelock==3.13.1
fsspec==2024.3.1
huggingface-hub==0.21.4
icecream==2.1.3
idna==3.6
Jinja2==3.1.3
MarkupSafe==2.1.5
mpmath==1.3.0
mysqlclient==2.2.4
networkx==3.2.1
numpy==1.26.4
packaging==24.0
pillow==10.2.0
Pygments==2.17.2
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
safetensors==0.4.2
six==1.16.0
sqlparse==0.4.4
sympy==1.12
tokenizers==0.15.2
torch==2.2.1+cu121
torchaudio==2.2.1+cu121
torchvision==0.17.1+cu121
tqdm==4.66.2
transformers==4.38.2
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
